Medium lint: ClassVar dicts, hoist stdlib imports, fix loop-var shadowing, unused locals
- F841: drop unused 'with...as' target and a dead assignment.
- RUF012: annotate class-level lookup dicts as ClassVar.
- PLC0415: hoist functools.reduce / math.gcd to module top.
- PLW2901: stop overwriting loop variables (introduce caption/matlabel/col/symbol).
- C416/RUF046/SIM/etc.: ruff safe + behaviour-preserving auto-fixes.
Full fast suite (2904) green.

+++ rtt/library/comma_forms.py
@@ -21,7 +21,7 @@ def _prime_limit(comma) -> int:
 
 
 def _as_matrix(rows) -> Matrix:
-    return tuple(tuple(int(round(x)) for x in r) for r in rows)
+    return tuple(tuple(round(x) for x in r) for r in rows)
 
 
 def positive_ratio_ca(matrix: Matrix, jip_octaves) -> Matrix:
